+++ tests/include/x509_chain_cs_2_static_dh/authentication_credentials_x5chain_cs_2_static_dh.h
@@ -1,0 +1,59 @@
+/**
+ * \file    authentication_credentials_x5chain_cs_2_static_dh.h
+ * \author  Kamil Kielbasa
+ * \brief   Example implementation of authentication credentials callbacks
+ *          for X.509 chain authentication method for cipher suite 2 with
+ *          static DH keys.
+ * \version 0.4
+ * \date    2024-01-01
+ * 
+ * \copyright Copyright (c) 2024
+ * 
+ */
+
+/* Header guard ------------------------------------------------------------ */
+#ifndef AUTHENTICATION_CREDENTIALS_X5CHAIN_CS_2_STATIC_DH_H
+#define AUTHENTICATION_CREDENTIALS_X5CHAIN_CS_2_STATIC_DH_H
+
+/* Include files ----------------------------------------------------------- */
+
+/* Standard library headers: */
+#include <stdint.h>
+#include <stddef.h>
+
+/* EDHOC header: */
+#include "edhoc_credentials.h"
+
+/* Defines ----------------------------------------------------------------- */
+/* Types and type definitions ---------------------------------------------- */
+/* Module interface variables and constants -------------------------------- */
+/* Extern variables and constant declarations ------------------------------ */
+/* Module interface function declarations ---------------------------------- */
+
+/**
+ * \brief Authentication credentials fetch callback for initiator.
+ */
+int auth_cred_fetch_init_x5chain_cs_2_static_dh(
+	void *user_context, struct edhoc_auth_creds *credentials);
+
+/**
+ * \brief Authentication credentials fetch callback for responder.
+ */
+int auth_cred_fetch_resp_x5chain_cs_2_static_dh(
+	void *user_context, struct edhoc_auth_creds *credentials);
+
+/**
+ * \brief Authentication credentials verify callback for initiator.
+ */
+int auth_cred_verify_init_x5chain_cs_2_static_dh(
+	void *user_context, struct edhoc_auth_creds *credentials,
+	const uint8_t **public_key_reference, size_t *public_key_length);
+
+/**
+ * \brief Authentication credentials verify callback for responder.
+ */
+int auth_cred_verify_resp_x5chain_cs_2_static_dh(
+	void *user_context, struct edhoc_auth_creds *credentials,
+	const uint8_t **public_key_reference, size_t *public_key_length);
+
+#endif /* AUTHENTICATION_CREDENTIALS_X5CHAIN_CS_2_STATIC_DH_H */