+++ tests/src/tests.c
@@ -19,6 +19,7 @@
 #include "x509_chain_cs_0/test_edhoc_handshake_x5chain_cs_0.h"
 #include "x509_chain_cs_2/test_edhoc_handshake_x5chain_cs_2.h"
 #include "x509_chain_cs_2/test_edhoc_handshake_x5chain_cs_2_ead.h"
+#include "x509_chain_cs_2_static_dh/test_edhoc_handshake_x5chain_cs_2_static_dh_ead.h"
 #include "x509_hash_cs_2/test_edhoc_handshake_x5t_cs_2_ead.h"
 #include "edhoc_trace_2/test_edhoc_handshake_2.h"
 #include "error_message/test_edhoc_error_message.h"
@@ -138,6 +139,10 @@ int main(void)
 	printf("\n");
 	printf("test_edhoc_handshake_x5chain_cs_2_single_cert_e2e_multiple_ead_tokens:\n");
 	test_edhoc_handshake_x5chain_cs_2_single_cert_e2e_multiple_ead_tokens();
+
+	printf("\n");
+	printf("test_edhoc_handshake_x5chain_cs_2_static_dh_keys_ead_e2e:\n");
+	test_edhoc_handshake_x5chain_cs_2_static_dh_keys_ead_e2e();
 
 	printf("\n");
 	printf("test_edhoc_handshake_x5t_cs_2_e2e_single_ead_token:\n");